Remove unused logging interfaces. Use logger instead of console.log everywhere.
