+++ shared/providers.ts
@@ -3,6 +3,7 @@ import { shareReplay } from 'rxjs/operators'
 import * as sourcegraph from 'sourcegraph'
 import { impreciseBadge } from './badges'
 import { LanguageSpec } from './language-specs/spec'
+import { Logger } from './logging'
 import { createProviders as createLSIFProviders } from './lsif/providers'
 import { createProviders as createSearchProviders } from './search/providers'
 import { noopAsyncGenerator, observableFromAsyncIterator } from './util/ix'
@@ -97,9 +98,10 @@ export class NoopProviderWrapper implements ProviderWrapper {
  * @param languageSpec The language spec used to provide search-based code intelligence.
  */
 export function createProviderWrapper(
-    languageSpec: LanguageSpec
+    languageSpec: LanguageSpec,
+    logger: Logger
 ): ProviderWrapper {
-    const lsifProviders = createLSIFProviders()
+    const lsifProviders = createLSIFProviders(logger)
     const searchProviders = createSearchProviders(languageSpec)
 
     return {
